requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20
pdfplumber==0.11.4
httpx==0.28.1
jinja2==3.1.5
python-dotenv==1.0.1
sqlalchemy==2.0.25
pyodbc==5.1.0
python-docx>=1.0.0
openpyxl>=3.1.0